Updated Builder-class
